Get aboutToQuit() signal from correct object
Since switching to QApplication in commit
4dd07cc0de266f4e9ae85a7e660ea8ed25fba811 the aboutToQuit() signal comes
directly from the app object itself.

This of course needs to be connected to the proper slots so the app can
release its connection to the hardware receiver correctly. Otherwise the
app may not get access to the receiver the next time it is started.